@@ -12,23 +12,31 @@
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
     "@mui/icons-material": "^5.11.0",
-    "@mui/material": "^5.11.3",
+    "@mui/material": "^5.11.8",
     "@mui/styled-engine-sc": "^5.11.0",
-    "@next/font": "^13.1.1",
+    "@next/font": "^13.1.6",
     "@react-spring/web": "^9.6.1",
+    "ky": "^0.33.2",
     "lodash": "^4.17.21",
-    "next": "13.1.1",
+    "mobx": "^6.8.0",
+    "mobx-react": "^7.6.0",
+    "moment": "^2.29.4",
+    "moment-locales-webpack-plugin": "^1.2.0",
+    "moment-timezone-data-webpack-plugin": "^1.5.1",
+    "next": "13.1.6",
     "node-sass": "^7.0.3",
     "query-string": "^8.1.0",
     "react": "18.2.0",
     "react-auth-code-input": "^3.2.1",
     "react-dom": "18.2.0",
-    "sass": "^1.57.1",
+    "sass": "^1.58.1",
     "styled-components": "^5.3.6",
-    "typescript": "^4.9.4"
+    "swiper": "^9.0.5",
+    "typescript": "^4.9.5",
+    "zustand": "^4.3.3"
   },
   "devDependencies": {
-    "eslint": "8.31.0",
-    "eslint-config-next": "13.1.1"
+    "eslint": "8.34.0",
+    "eslint-config-next": "13.1.6"
   }
 }
