@@ -20,6 +20,9 @@ import '../styles/footer.scss'
 
 import { createTheme, ThemeProvider } from '@mui/material/styles';
 
+import { Provider } from 'mobx-react';
+import { itemsStoreNew as stores} from '../components/items-store.js';
+
 const theme = createTheme({
   palette: {
     primary: {
@@ -30,9 +33,11 @@ const theme = createTheme({
 
 function MyApp({ Component, pageProps }) {
   return (
-    <ThemeProvider theme={theme}>
-      <Component {...pageProps} />
-    </ThemeProvider>
+    <Provider { ...stores }>
+      <ThemeProvider theme={theme}>
+        <Component {...pageProps} />
+      </ThemeProvider>
+    </Provider>
   )
 }
 
